refactor: reduce unneeded exports
